Update winit and glutin
Which will allow much more DPI-independent goodness!

+++ src/conf.rs
@@ -46,14 +46,14 @@ pub enum FullscreenType {
 ///     max_height: 0,
 /// }
 /// ```
-#[derive(Debug, Copy, Clone, SmartDefault, Serialize, Deserialize, PartialEq, Eq)]
+#[derive(Debug, Copy, Clone, SmartDefault, Serialize, Deserialize, PartialEq)]
 pub struct WindowMode {
     /// Window width
-    #[default = r#"800"#]
-    pub width: u32,
+    #[default = r#"800.0"#]
+    pub width: f64,
     /// Window height
-    #[default = r#"600"#]
-    pub height: u32,
+    #[default = r#"600.0"#]
+    pub height: f64,
     /// Whether or not to maximize the window
     #[default = r#"false"#]
     pub maximized: bool,
@@ -64,22 +64,22 @@ pub struct WindowMode {
     #[default = r#"false"#]
     pub borderless: bool,
     /// Minimum width for resizable windows; 0 means no limit
-    #[default = r#"0"#]
-    pub min_width: u32,
+    #[default = r#"0.0"#]
+    pub min_width: f64,
     /// Minimum height for resizable windows; 0 means no limit
-    #[default = r#"0"#]
-    pub min_height: u32,
+    #[default = r#"0.0"#]
+    pub min_height: f64,
     /// Maximum width for resizable windows; 0 means no limit
-    #[default = r#"0"#]
-    pub max_width: u32,
+    #[default = r#"0.0"#]
+    pub max_width: f64,
     /// Maximum height for resizable windows; 0 means no limit
-    #[default = r#"0"#]
-    pub max_height: u32,
+    #[default = r#"0.0"#]
+    pub max_height: f64,
 }
 
 impl WindowMode {
     /// Set default window size, or screen resolution in true fullscreen mode
-    pub fn dimensions(mut self, width: u32, height: u32) -> Self {
+    pub fn dimensions(mut self, width: f64, height: f64) -> Self {
         self.width = width;
         self.height = height;
         self
@@ -104,14 +104,14 @@ impl WindowMode {
     }
 
     /// Set minimum window dimensions for windowed mode
-    pub fn min_dimensions(mut self, width: u32, height: u32) -> Self {
+    pub fn min_dimensions(mut self, width: f64, height: f64) -> Self {
         self.min_width = width;
         self.min_height = height;
         self
     }
 
     /// Set maximum window dimensions for windowed mode
-    pub fn max_dimensions(mut self, width: u32, height: u32) -> Self {
+    pub fn max_dimensions(mut self, width: f64, height: f64) -> Self {
         self.max_width = width;
         self.max_height = height;
         self
@@ -132,7 +132,7 @@ impl WindowMode {
 ///     samples: NumSamples::One,
 /// }
 /// ```
-#[derive(Debug, Clone, SmartDefault, Serialize, Deserialize, PartialEq, Eq)]
+#[derive(Debug, Clone, SmartDefault, Serialize, Deserialize, PartialEq)]
 pub struct WindowSetup {
     /// The window title.
     #[default = r#""An easy, good game".to_owned()"#]
